+++ programs/cp-amm/src/instructions/initialize_pool/ix_initialize_pool_with_dynamic_config.rs
@@ -242,7 +242,6 @@ pub fn handle_initialize_pool_with_dynamic_config<'c: 'info, 'info>(
         ctx.accounts.token_a_vault.key(),
         ctx.accounts.token_b_vault.key(),
         alpha_vault,
-        config.pool_creator_authority,
         sqrt_min_price,
         sqrt_max_price,
         sqrt_price,